# packages/server/types/api.ts
import type { components } from "./api.gen";
import { webhook } from "@line/bot-sdk";

export type WebhookStreamObject = components["schemas"]["WebhookStreamObject"];
export type ErrorObject = components["schemas"]["ErrorObject"];